Day-7/hackathon/regulation-aware-site-auditor/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.8.3
python-dotenv==1.0.0
validators==0.22.0
pydantic==2.9.2
python-multipart==0.0.6
cors==1.0.1
fastapi-cors==0.0.6 